% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{get_checkouts}
\alias{get_checkouts}
\title{Obtain the checkout data matching a given location substring}
\usage{
get_checkouts(location)
}
\arguments{
\item{location}{A location substring that filters the reed_checkouts with a location string containing the substring.}
}
\value{
A checkouts data frame.
}
\description{
Returns a filtered reed_checkouts dataset matching that location substring
}
\examples{

# Get PARC checkouts
PARC_checkouts <- get_checkouts(location = "PARC")

# Get IMC checkouts
IMC_checkouts <- get_checkouts(location = "IMC")

# Get musical score checkouts
score_checkouts <- get_checkouts(location = "score")

}
