% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reed_checkouts.R
\docType{data}
\name{reed_checkouts}
\alias{reed_checkouts}
\title{Reed College Library Checkouts (2018-2020)}
\format{
A data frame with Reed Checkout Data
\itemize{
\item{\code{Title}}{ title (Book title)}
}
}
\source{
Data provided by Mahria Lebow
}
\usage{
data(reed_checkouts)
}
\description{
Check out data from the Reed College Library from 8/1/2018 to 7/30/20.
}
\keyword{datasets}
